Add init command and migrations
This commit adds a new `init` command which will initialize a SQLite
database with the schema. Initializing the database is required prior to
running `import`. The database schema is managed with sqlite-migrate
which is kind of new, but seems to work well.

So the process for working with warcdb is to:

```bash
$ warcdb init warc.db
$ warcdb import warc.db google.warc.gz
```

Then if you you want to update warcdb and apply migrations you can:

```
$ pip install --upgrade warcdb
$ warcdb migrate warc.db
```

Closes #12